refactor(release): keep portable smoke within capsule runtime
