+++ docs/adr/KF-ADR-019f86da-4f90-7b6b-ae6d-76cea57487f2.md
@@ -6,7 +6,7 @@ decision_status: accepted
 implementation_status: implemented
 implementation_prs: [https://github.com/kungfu-systems/kungfu/pull/731, https://github.com/kungfu-systems/kungfu/pull/737, https://github.com/kungfu-systems/kungfu/pull/1335, https://github.com/kungfu-systems/kungfu/pull/1340, https://github.com/kungfu-systems/kungfu/pull/1342, https://github.com/kungfu-systems/kungfu/pull/1347, https://github.com/kungfu-systems/kungfu/pull/1351, https://github.com/kungfu-systems/kungfu/pull/1514, https://github.com/kungfu-systems/kungfu/pull/1528, https://github.com/kungfu-systems/kungfu/pull/1582, https://github.com/kungfu-systems/kungfu/pull/1599, https://github.com/kungfu-systems/kungfu/pull/1603, https://github.com/kungfu-systems/kungfu/pull/1606, https://github.com/kungfu-systems/kungfu/pull/1607, https://github.com/kungfu-systems/kungfu/pull/1612, https://github.com/kungfu-systems/kungfu/pull/1618, https://github.com/kungfu-systems/kungfu/pull/1626, https://github.com/kungfu-systems/kungfu/pull/1642, https://github.com/kungfu-systems/kungfu/pull/1931, https://github.com/kungfu-systems/kungfu/pull/1941, https://github.com/kungfu-systems/kungfu/pull/1955, https://github.com/kungfu-systems/kungfu/pull/2001, https://github.com/kungfu-systems/kungfu/pull/2016, https://github.com/kungfu-systems/kungfu/pull/2020, https://github.com/kungfu-systems/kungfu/pull/2210, https://github.com/kungfu-systems/kungfu/pull/2241, https://github.com/kungfu-systems/kungfu/pull/2251, https://github.com/kungfu-systems/kungfu/pull/2376, https://github.com/kungfu-systems/kungfu/pull/2444, https://github.com/kungfu-systems/kungfu/pull/2454, https://github.com/kungfu-systems/kungfu/pull/2457, https://github.com/kungfu-systems/kungfu/pull/2471, https://github.com/kungfu-systems/kungfu/pull/2501, https://github.com/kungfu-systems/kungfu/pull/2549, https://github.com/kungfu-systems/kungfu/pull/2554, https://github.com/kungfu-systems/kungfu/pull/2597]
 closure_pr: https://github.com/kungfu-systems/kungfu/pull/731
-qualification_refs: [.buildchain/buildchain.toml, .buildchain/auditable-demo.json, .github/actions/require-alpha-preflight/alpha-macos-overflow.mjs, .github/workflows/affected-native-pr.yml, .github/workflows/build.yml, .github/workflows/dev-alpha-candidate-patrol.yml, .github/workflows/dev-delivery-warrant-terminal.yml, .github/workflows/dev-pr-auto-merge.yml, .github/workflows/release-new-version.yml, .github/workflows/stable-candidate-patrol.yml, docs/qualification/alpha-local-publication-debug.declaration.json, docs/qualification/alpha-local-publication-debug.md, docs/qualification/alpha-release-latency.contract.json, docs/qualification/alpha-ruleset.contract.json, docs/qualification/auditable-demo-artifact-pipeline.md, docs/qualification/stable-release-continuation.contract.json, docs/qualification/stable-ruleset.contract.json, docs/release-promotion-rehearsal.contract.json, framework/core/tests/python/test_peer_lifecycle.py, framework/core/tests/python/windows_continuity_sentinel.py, framework/release/alpha-local-publication-debug/index.mjs, framework/release/alpha-local-publication-debug/portable-smoke.mjs, framework/release/publication-control-plane.mjs, framework/release/publication-surfaces.json, framework/version-line/check-version-line-authority.mjs, framework/version-line/version-line-authority.json, framework/version-line/version-line-authority.mjs, framework/version-line/version-line-projections.json, product/scripts/dist.mjs, product/scripts/dist.test.mjs, scripts/adr-release-gate.test.mjs, scripts/affected-native-proof.mjs, scripts/affected-native-proof.test.mjs, scripts/alpha-local-publication-debug.test.mjs, scripts/alpha-macos-overflow.test.mjs, scripts/auditable-demo-workflow.test.mjs, scripts/release-promotion-rehearsal.test.mjs, scripts/alpha-promotion-preflight.mjs, scripts/alpha-promotion-preflight.test.mjs, scripts/alpha-release-timeline.test.mjs, scripts/alpha-release-history.test.mjs, scripts/alpha-cache-evidence.test.mjs, scripts/alpha-ruleset.test.mjs, scripts/dev-alpha-candidate-patrol.test.mjs, scripts/dev-delivery-warrant-input.mjs, scripts/dev-delivery-warrant-input.test.mjs, scripts/dev-pr-auto-merge.test.mjs, scripts/stable-candidate-patrol.test.mjs, scripts/release-publication-control-plane.test.mjs, scripts/source-acceptance.mjs, scripts/version-line-authority.test.mjs]
+qualification_refs: [.buildchain/buildchain.toml, .buildchain/auditable-demo.json, .github/actions/require-alpha-preflight/alpha-macos-overflow.mjs, .github/workflows/affected-native-pr.yml, .github/workflows/build.yml, .github/workflows/dev-alpha-candidate-patrol.yml, .github/workflows/dev-delivery-warrant-terminal.yml, .github/workflows/dev-pr-auto-merge.yml, .github/workflows/release-new-version.yml, .github/workflows/stable-candidate-patrol.yml, docs/qualification/alpha-local-publication-debug.declaration.json, docs/qualification/alpha-local-publication-debug.md, docs/qualification/alpha-release-latency.contract.json, docs/qualification/alpha-ruleset.contract.json, docs/qualification/auditable-demo-artifact-pipeline.md, docs/qualification/stable-release-continuation.contract.json, docs/qualification/stable-ruleset.contract.json, docs/release-promotion-rehearsal.contract.json, framework/core/tests/python/test_peer_lifecycle.py, framework/core/tests/python/windows_continuity_sentinel.py, framework/release/alpha-local-publication-debug/index.mjs, framework/release/publication-control-plane.mjs, framework/release/publication-surfaces.json, framework/version-line/check-version-line-authority.mjs, framework/version-line/version-line-authority.json, framework/version-line/version-line-authority.mjs, framework/version-line/version-line-projections.json, product/scripts/dist.mjs, product/scripts/dist.test.mjs, scripts/adr-release-gate.test.mjs, scripts/affected-native-proof.mjs, scripts/affected-native-proof.test.mjs, scripts/alpha-local-publication-debug.test.mjs, scripts/alpha-macos-overflow.test.mjs, scripts/auditable-demo-workflow.test.mjs, scripts/release-promotion-rehearsal.test.mjs, scripts/alpha-promotion-preflight.mjs, scripts/alpha-promotion-preflight.test.mjs, scripts/alpha-release-timeline.test.mjs, scripts/alpha-release-history.test.mjs, scripts/alpha-cache-evidence.test.mjs, scripts/alpha-ruleset.test.mjs, scripts/dev-alpha-candidate-patrol.test.mjs, scripts/dev-delivery-warrant-input.mjs, scripts/dev-delivery-warrant-input.test.mjs, scripts/dev-pr-auto-merge.test.mjs, scripts/stable-candidate-patrol.test.mjs, scripts/release-publication-control-plane.test.mjs, scripts/source-acceptance.mjs, scripts/version-line-authority.test.mjs]
 review_state: self-reviewed
 sensitivity: public
 sources: [local-files, user-decision]
