+++ docs/architecture/adr-map.md
@@ -211,7 +211,7 @@ review`; evidence is the number of declared qualification references.
 
 | Key | Decision | Theme | Status | Evidence | Nearby (navigation only) |
 |---|---|---|---|---:|---|
-| KF · 87f2 | [Buildchain promotion is the settlement boundary for ADR implementation truth](../adr/KF-ADR-019f86da-4f90-7b6b-ae6d-76cea57487f2.md) | buildchain-adr-release-admissibility | accepted / implemented / self-reviewed | 51 | [KF · e848](../adr/KF-ADR-019f8cc4-e796-7b0e-9a16-50c08614e848.md) (shared terms: boundary, release)<br>[SHIFU · 4d50](../adr/SHIFU-ADR-019f86da-4f90-7eac-ad5f-db132ae04d50.md) (shared terms: release)<br>[SHIFU · 4453](../adr/SHIFU-ADR-019f86da-4f90-7885-9597-bfdcb4fd4453.md) (shared terms: promotion)<br>[KF · c2aa](../adr/KF-ADR-019fbd57-00fe-74a6-b2e7-11579923c2aa.md) (shared terms: release) |
+| KF · 87f2 | [Buildchain promotion is the settlement boundary for ADR implementation truth](../adr/KF-ADR-019f86da-4f90-7b6b-ae6d-76cea57487f2.md) | buildchain-adr-release-admissibility | accepted / implemented / self-reviewed | 50 | [KF · e848](../adr/KF-ADR-019f8cc4-e796-7b0e-9a16-50c08614e848.md) (shared terms: boundary, release)<br>[SHIFU · 4d50](../adr/SHIFU-ADR-019f86da-4f90-7eac-ad5f-db132ae04d50.md) (shared terms: release)<br>[SHIFU · 4453](../adr/SHIFU-ADR-019f86da-4f90-7885-9597-bfdcb4fd4453.md) (shared terms: promotion)<br>[KF · c2aa](../adr/KF-ADR-019fbd57-00fe-74a6-b2e7-11579923c2aa.md) (shared terms: release) |
 | KF · e848 | [Kungfu UNGFU source signature and public-release evidence boundary](../adr/KF-ADR-019f8cc4-e796-7b0e-9a16-50c08614e848.md) | kungfu-ungfu-public-use | accepted / staged / self-reviewed | 15 | [KF · 87f2](../adr/KF-ADR-019f86da-4f90-7b6b-ae6d-76cea57487f2.md) (shared terms: boundary, release)<br>[SHIFU · 4d50](../adr/SHIFU-ADR-019f86da-4f90-7eac-ad5f-db132ae04d50.md) (shared terms: release)<br>[KF · c2aa](../adr/KF-ADR-019fbd57-00fe-74a6-b2e7-11579923c2aa.md) (shared terms: release)<br>[KF · c239](../adr/KF-ADR-019fabb5-62a0-7b8d-8f8d-6505efdbc239.md) (shared terms: release) |
 | KF · c239 | [Product Release Cut is the exact updater identity](../adr/KF-ADR-019fabb5-62a0-7b8d-8f8d-6505efdbc239.md) | product-release-cut-updater | accepted / staged / self-reviewed | 10 | [KF · c2aa](../adr/KF-ADR-019fbd57-00fe-74a6-b2e7-11579923c2aa.md) (shared terms: cut, product, release)<br>[KF · 87f2](../adr/KF-ADR-019f86da-4f90-7b6b-ae6d-76cea57487f2.md) (shared terms: release)<br>[SHIFU · 4d50](../adr/SHIFU-ADR-019f86da-4f90-7eac-ad5f-db132ae04d50.md) (shared terms: release)<br>[KF · e848](../adr/KF-ADR-019f8cc4-e796-7b0e-9a16-50c08614e848.md) (shared terms: release) |
 | KF · c2aa | [Product Release Cut history is portable independently of caches](../adr/KF-ADR-019fbd57-00fe-74a6-b2e7-11579923c2aa.md) | durable-history-data-protection | accepted / staged / self-reviewed | 4 | [KF · c239](../adr/KF-ADR-019fabb5-62a0-7b8d-8f8d-6505efdbc239.md) (shared terms: cut, product, release)<br>[KF · 87f2](../adr/KF-ADR-019f86da-4f90-7b6b-ae6d-76cea57487f2.md) (shared terms: release)<br>[SHIFU · 4d50](../adr/SHIFU-ADR-019f86da-4f90-7eac-ad5f-db132ae04d50.md) (shared terms: release)<br>[KF · e848](../adr/KF-ADR-019f8cc4-e796-7b0e-9a16-50c08614e848.md) (shared terms: release) |
