+++ package.json
@@ -239,7 +239,7 @@
     "check:release-publication-control-plane": "node scripts/require-shifu.mjs check:release-publication-control-plane && node framework/release/publication-control-plane.mjs check",
     "test:release-publication-control-plane": "node scripts/require-shifu.mjs test:release-publication-control-plane && node --test scripts/release-publication-control-plane.test.mjs",
     "alpha:publication:debug": "node scripts/require-shifu.mjs alpha:publication:debug && node framework/release/alpha-local-publication-debug/index.mjs",
-    "alpha:publication:debug:portable-smoke": "node scripts/require-shifu.mjs alpha:publication:debug:portable-smoke && node framework/release/alpha-local-publication-debug/portable-smoke.mjs",
+    "alpha:publication:debug:portable-smoke": "node scripts/require-shifu.mjs alpha:publication:debug:portable-smoke && node framework/release/alpha-local-publication-debug/index.mjs portable-smoke",
     "test:alpha-publication-debug": "node scripts/require-shifu.mjs test:alpha-publication-debug && node --test scripts/alpha-local-publication-debug.test.mjs",
     "release:promotion:rehearse": "node scripts/require-shifu.mjs release:promotion:rehearse && node scripts/release-promotion-rehearsal.mjs",
     "release:probe:platform": "node scripts/require-shifu.mjs release:probe:platform && node scripts/probe-release-platform.mjs",
